Sync atmospheric heater "reset" gizmo (#495)
`Comp_AtmosphericHeater` is a subtype of `CompTempControl`, but has the `CompGetGizmosExtra` method rewritten. The +/- 1/10 degrees change was working fine, as it called the synced method `InterfaceChangeTargetTemperature_NewTemp`. However, this was not the case for the reset gizmo, as it set the temperature directly. I've synced the reset gizmo for the atmospheric heater itself, so it should work fine now.

Also moved the `CompTempControl` gizmo from `SyncDelegates` to `SyncMethods`, right next to where the other sync method for it was registered. I believe it should have been there to begin with, and I must have made a mistake while working through all the new and changed gizmos in 1.5.