Fix void monolith float menu causing desync with `CompProximityLetter` (#499)
`CompProximityLetter` (void monolith, ancient mechanitor exostrider) sends a letter when a player pawn gets close to it. This is disabled (`CompProximityLetter.letterSent` is set to true) when the void monolith is automatically activated, or when a pawn is ordered to interact with it.

When a pawn interacts with the monolith by using the gizmo, the method disabling the letter is synced. However, when using float menu to do it the letter is only disabled for the player who ordered it, as the method is not synced (instead, the interaction is synced by `Pawn_JobTracker.TryTakeOrderedJob` call). It then displays the letter for everyone besides the player who ordered the interaction, causing desync (new ID assigned, causing desync trace mismatch).

The fix was rather simple and required syncing the method the float menu calls when clicked. Investigating what caused the desync took much longer.

I've also moved the void monolith sync methods from `SyncMethods` file to `SyncDelegates` to keep them categorized, specifically by keeping them together with the newly added sync delegate.