Fix ping always showing 0ms when connecting without using LiteNetLib (#529)
The root of the issue is that the keepAliveTimer was only started once
when creating the ServerPlayer and then always just `Reset()` which does
not start it again. The timer was only ever started once and then reset
while not having counted any time.

To test this remove ServerPlayingState:194 (if connection is LNL: return)
and OnNetworkLatencyUpdate in NetworkingLiteNet

As a drive-by improvement I added some minor comments nearby.