Prevent duplicate monolith generation in multifaction (#563)
Due to changes in #508, an Exorider is now spawned on every newly generated multifaction map. However, this also caused a Monolith to be generated each time, resulting in multiple Monoliths appearing and causing issues.

This update adds a Harmony patch to GenStep_Monolith.GenerateMonolith to skip Monolith generation if one already exists.