Fix scenario research lost on reconnect (#588)
Fixes #585 .

ScenPart_StartingResearch.PostGameStart was only executed on the local machine of the faction creator.
As a result, the host/save did not reflect this change properly.

I’ve opened an issue #587 to improve the overall handling of ScenParts. I believe there’s more to gain by reviewing the other available part scenParts. I’ve added this as a TODO for future work.