Fix camp is not abandoned correctly when reforming the caravan and relating problems (#603)
* Refactor camp creation

* Fixing camp not abandoning

Fixing camp not abandoning on caravan reform in multiplayer and disabled this patch in singleplayer, because it happend there as well.

* Map removal out of bounce error

In map deinitialize and removal process it gets removed from game.maps.
If it is the last map added then this patch throws a out of bounce exception in the removal process due to calling Find.CurrentMap while currentMapIndex points to an invalid index.

* Fix warning in camp creation

The patch could not find a faction for the current tile and threw a warning. This commit includes caravans into the search for a faction at the tile

* Clean up