More forgiving PlayerList packet handling (#608)
* Add some more metadata to errors when reading a packet fails

This should add some extra hints when diagnosing problems with packets.
Limited the display to only the first 128 bytes so that big packets
won't cause issues.

* Make PlayerListAction.Latencies handler resistant to player mismatches

* Add diagnostics for unexpected scenarios in PlayerList packet handling

* Make sure the player list is sent before any other related packets.

As soon as a connection's state becomes Playing, player list updates can
be sent to it. However, sending a player list update before sending the
player list itself is wrong, so make sure to send it before other
packets. The timing required to trigger this seems improbable, but I did
see issues that could be related to this.

* Log notifications and some timing info on the client-side when loading