Sync "Drop pod raid at location" (#578)
* Sync new 1.6 debug option Drop pod raid at... #556

This PR fixes #556 and synchronizes the last step in "Drop pod raid at location..." debug option that utilizes the `Targeter.BeginTarget` method. This method isn't used elsewhere in the debug system, but it is invoked by certain other systems. It basically takes a action and invokes it on mouse click.

I implemented the fix in a similiar fession to the other `DebugSource`.

The main challenge I faced was ensuring that my prefix is executed _only_ when `Targeter.BeginTarget` is called from the `DebugActionsIncidents.ExecuteDropPodRaidAtLocation` method.
To achieve this, I used a bit of reflection to check where the action was created. I'm not entirely sure if this is the optimal approach. Feedback on better alternatives is welcome.

* Track current executing command type #556

Tracking the currently executing command type allows me to run the Targeter.BeginTargeting prefix only when it's executed within DebugToolCmd.

Additionally, this change made some of the logic that checks whether a cmd is currently running redundant, so I removed it."