Gravship Basic Behavior (#631)
* - Setup Gravship Travel session
- Synced up some of the basic for traveliing, works with no desync if temp terrain and vtr are off.
- Fixed some issues with map swapping.

* - Synced landing animation
- Reenabled vtr
- Synced dev tools for Pilot Console and Breakdownable.

* Remove delegate that isnt actually a delegate :)

* Potentially fixed some issues with closing session

* Added logs, synced takeoff

* Now freeze clients during visual processes that depend on the machine specs.
Syncing landing outcome defs
Cleaned up some things, remove unnecessary sync methods.

* Fixed cancelling on world map

* - Can now move the ship landing position
- Added serialization for Designator_MoveGravship
- Added some scribe data to GravshipTravelSession - still needs work
- Moved Cancelling while in world map so that it is only checked and called when cancelled, not selected.

* Fixes Bugs When Cancel Gravship Launch On Worldmap

A recent commit introduced a change that causes session cancellation to result in a desync and leaving the session open.

* Clears Tile Picker For Non Issuer In Gravship Launch

This commit fixes an incorrect call to TilePicker.StopTargeting on non-Issuers when the gravship has a valid destination tile selected and is about to start.

Issuers clean up the TargetPicker after selecting a valid tile and preparing to launch the gravship.

This commit introduces the same behavior for non-Issuers.

* Fix: Cancel Launch On World Map Not Closing For Non Issuer

Refactor to use the original method for StopTargeting.

I spent quite a while trying to find what in the source code causes the map to not close for the non-issuer. I found that, for the issuer, the wantedMode is set to None and for the issuer to planet, but I couldn’t find where this happens in the source code. So I just set it manually.

* Closing Gravship Travel Session Correctly

The gravship travel session was not closed correctly, so it was saved and caused errors on the next load.

* Adding Namespace To GravshipTravekSession.cs

* Simplifies Gravship Sessions Handling

This commit fixes loading session errors, and joining a client during takeoff, landing, or an active session is now possible.

When a user joins, the current session is canceled. This works well for the takeoff session. In the landing session, however, when the landing position is selected, the session is not recreated. As a result, the time can still be controlled. I think this is fine for now, but I’ll note it down for future change.

We now use separate takeoff and landing sessions.
The world session has been removed.
Exposable sessions have also been removed.

* Fix Gravship TakeOff Session with Anchor

The take-off session wasn’t opened due to a different code path, so our prefix for opening the session was never executed.

In other words, the map destruction confirmation window isn’t displayed because the map itself isn’t being destroyed with a anchor.

* Fix Gravship Landing on Anchored Map

Flying back to an anchored map is handled differently.

Fixes:

Map could not be found when opening a session due to the marker being initialized differently.

Marker could not be synced correctly because it was not spawned or did not have a parent map.

Deselects the gravship designator if another client selects a landing position.

* Improves Syncing Of Landing Confirmation Button

Removes half copied WorldComponentOnGUI and syncs the confirmation button in another way.

* Implements Abort Button On Anchored Maps

This commit implements the abort button which redirects you to a random map nearby.

I did experience really rare Rand problems, where a other abort map was picked between clients.

* Refactor for redundancy

---------

Co-authored-by: Reznal <cwilksch@gmail.com>
Co-authored-by: Meru <notfood@users.noreply.github.com>