Use built-in constants for tick values (#649)
I've replaced a few of the constant values (that I could catch) used in the mod with references to the in-game constants.

I've done this primarily to make the intentions behind those values clearer. Using `GenTicks.TicksPerRealSecond` is more obvious that the value refers to amount of ticks in a second, rather than just using the value of `60`. Likewise, using `GenDate.TicksPerDay` is much easier to understand than a value of `60000`, or even `2500 * 24`.

The only "change" on top is changing `1 / 60f` to `1f / GenTicks.TicksPerRealSecond`, as we need one of those values to be a float (or we'll end up with integer division).

This will not change anything gameplay-wise, as all those constants will be computet at compile time and end up the same in the compiled code. This will only affect the source code.