Fixes Random Desyncs Due To Statue Exception And NextThingId Missmatch (#632)
* Fixes New Statue Exception And NextThingId Missmatch

1.6 introduces Statues, which causes problems.
This fixes two issues with the new method InitFakePawn:

1. Pawn name sync issue: The pawn name was being synced because the getter was called before the pawn was fully initialized. At that point, the pawn was not set up correctly, which caused syncing errors due to the missing PawnKindDef.

2. Desyncs due to UniqueIdsPatch: This was the main reason I investigated the issue. Many desyncs occurred because UniqueIdsPatch generates negative thing IDs if certain values are set in a specific way. These values differ between client and host — specifically the Multiplayer.InInterface property.
The method InitFakePawn is called inside an ExecuteWhenFinished event. On the host, this event runs as an async long event, while on the client it runs as a synced event. As a result, the LongEventHandler.currentEvent field is different on host and client, which makes InInterface differ as well.
This causes the host to generate negative new thing IDs, while the clients generate positive ones. Since clients get ahead in nextThingIds, the next time a thing is generated a desync occurs.

* Small Refactor For Odyssey Multifaction Transpiler

* Change Pawn.Name To Use Pawn.nameInt instead

Calling the Pawn.Name getter in CompStatue.InitFakePawn causes issues because the pawn has not been initialized correctly at this point.

* Implementing Sokyrans proposal #632

Fix Problem 2 from #632 by adjusting the execution of the cleanup logic in UpdateCurrentAsynchronousEvent, as suggested by @SokyranTheDragon

* Refactor naming for CompStatue.InitFakePawn patch

* Revert Change For Syncing CompStatue.InitFakePawn